Make source links relative to OUT_DIR
When generating code, `cranelift-isle` records the path to the source
files. These paths were sometimes absolute, meaning it kept a trace of
the system it was run on.

These changes here add helper functions in `Files` to try and strip the
value of `$OUT_DIR` (if set) from the file names.