test-macros: Handle custom test attributes
This commit introduces the ability to pass a custom attribute to the
procedural macro via `#[wasmtime_test(with = "...")]`. This makes it
possible to use the macro in situations where the caller desires to use
a different test attribute such as `#[tokio::test]`.

Additionally this commit also introduces support for the test funcition
asyncness, which was previously ignored.

This change is mainly motivated by adding epoch interruption support to
Winch.