Remove feature validation from Wasmtime
Some of this originated in #917 but nowadays it shouldn't be necessary
to control proposals like this. Instead it's probably best nowadays to
throw configuration at `wasmparser` and use its definition of features
to determine whether constructs make sense or not. This reduces the
amount of bits and pieces Wasmtime has to do and avoids interactions
such as #9622.

Closes #9622