Change `VMArrayCallFunction` to be an opaque pointer
This pointer is technically not safe to be a function pointer because it
will get unloaded when the module is dropped. Additionally with Pulley
this isn't actually a native function pointer. This builds on #9630 to
prepare for future integration with Pulley to ensure that the number of
locations that have to deal with bytecode-vs-native-code are minimized.