Don't panic when instructions aren't in ISLE (#9631)
Instead return a codegen error which can be more easily caught,
displayed, and rendered. This is intended to help with enabling testing
of Pulley in the near future where Pulley doesn't implement most
instructions at this time and this should make the in-progress state of
the backend a bit easier to test.