Optimize a test exercising Cranelift's limits (#9637)
This is a test which generates a wasm function that exceeds Cranelift's
limits of virtual registers. Previously in debug mode this test took
upwards of 40 seconds to execute and now it takes around 5. The main
optimization here was to disable Cranelift optimizations. The test has
also been updated to generate the module with `wasm-encoder` instead of
`wast` to skip the text parsing step too.