pulley: Implement a new `br_table` instruction
This is intended to match WebAssembly's `br_table` and Cranelift's
version as well. This is implemented as a new `br_table32` opcode where
a 32-bit number of branch targets are encoded after `br_table32` all as
a `PcRelOffset`, a 32-bit offset. This helps bake in a more "macro
opcode" into the interpreter rather than a handful of more primitive
opcodes that would achieve the same result with loads/indirect
jumps/comparisons/etc.