Compile wasm builtins with the `tail` convention, not host
This commit updates how Cranelift compiles the shims that go from
WebAssembly to various libcall builtins (e.g. `memory.grow`). Previously
the small trampolines in Cranelift were compiled with the native host
calling convention and this commit instead changes them to the wasm
calling convention (e.g. "tail"). This helps keep caller/callee matched
with calling conventions in wasm code itself and shifts the boundary of
the wasm->host transition to the indirect call of the host function
pointer itself. The general hope is that this makes Pulley a bit easier
since it'll want to minimize handling of wasm->host transitions.