Update the host ABI of Wasmtime to return failure
This commit updates the "array call" ABI of Wasmtime used to transition
from wasm to the host to explicitly return a `bool` indicating whether
the call succeeded or not. Previously functions would implicitly unwind
via `longjmp` and thus no explicit checks were necessary. The burden of
unwinding is now placed on Cranelift-compiled code itself instead of the
caller.

There are a few pieces of rationale for this change:

* Primarily I was implementing initial integration of Pulley where the
  host `setjmp` and `longjmp` cannot be used to maintain the Pulley
  interpreter state. My initial plan for handling this was to handle
  traps a bit differently in Pulley where having direct access to
  whether a function trapped or not in the interpreter bytecode is
  easier to deal with.

* Additionally it's generally not safe to call `longjmp` from Rust as it
  will not run on-stack destructors. This is ok today in the sense that
  we shouldn't have these in Wasmtime, but directly returning to
  compiled code improves our safety story here a bit where we just won't
  have to deal with the problem of skipping destructors.

* In the future I'd like to move away from `setjmp` and `longjmp`
  entirely in the host to a Cranelift-native solution. This change would
  be required for such a migration anyway so it's my hope to make such a
  Cranelift-based implementation easier in the future. This might be
  necessary, for example, when implementing the `exception-handling`
  proposal for Wasmtime.

Functionally this commit does not remove all usages of
call-`longjmp`-from-Rust. Notably all libcalls and builtins still use
this helper in the trampolines generated in Rust. I plan on going
through the libcalls and updating their ABIs and signatures to reflect
this in follow-up commits. As a result a few preexisting functions that
should go away are marked `#[deprecated]` for internal use in this
commit. I'll be cleaning that up as follow-ups. For now this commit
handles the "hard part" of host functions by ensuring that the new
`bool` return value is plumbed in all the locations correctly.

prtest:full