pulley: Remove virtual sp offset
This fixes the output of the pulley cranelift backend with respect to
virtual stack offsets and adjustments for calls with stack parameters
and results. Previously the virtual stack pointer was updated but the
rest of the backend assumed the real stack pointer was updated. This
updates the pulley backend to match the other backends in this respects
and fixes a test on my "more complete pulley integration" branch.