Pin some runner images (#9717)
For the larger test/build jobs use some variables in the `*.js` files
calculating the matrix to pin the images to fixed versions of GitHub
actions images. For the `main.yml` file instead of changing everything
over to a pin only do the ones that are currently failing on the
`ubuntu-24.04` update. The hope is that there's only a few locations to
update pinned versions in the future, and we'll need to keep an eye on
CI warnings and such to know when to update these pins in the future.