Default to Pulley on cranelift/winch-unsupported platforms
This commit updates the selection of the default target used by a
`Config` to take into account the host architecture and possibly use
Pulley instead of the host target itself. The goal here is that when a
target isn't explicitly configured then Wasmtime is tasked with picking
a reasonable default to execute code with. If neither Winch nor
Cranelift has any support then the only way to possibly execute code
would be with Pulley, so in these situations pulley becomes the default
target.

The goal of this change is to make testing easier on 32-bit platforms
where there is no compiler support. This means we won't have to update
all tests to explicitly configure pulley on unsupported platforms.
Additionally this means that an eventual `wasmtime` executable for
32-bit platforms will use Pulley by default and won't need any extra
configuration.