pulley: Track faulting opcode in stack overflow better (#9735)
* pulley: Track faulting opcode in stack overflow better

This commit updates the pulley `Encode` trait to have a `WIDTH`
associated with it to be able to calculate the size of an instruction by
name rather than hard-coding instruction details in multiple locations.
This constant is procedurally generated per-instruction given the
definition of the instruction. This for now assumes there are no
variable-width instructions.

* Fix tests

* Fix feature'd build