Don't cancel workflows on Windows runners
Our CI has configuration such that on failure of any job it cancels the
entire workflow run. This means that if a failure is quickly encountered
it doesn't wait for the entire workflow to finish just to report the
failure, clearing out the queue of failing PRs faster. This doesn't work
well on Windows runners though where often the cancellation and marking
the job as failed race. The winner of the race is often the cancellation
which means that there's no listed failure in 100+ jobs which can be
frustrating.

Instead centralize this cancellation logic in one helper and
additionally add a condition where it doesn't cancel on Windows
platforms. That means that failures on windows won't be fail-fast, but
that's sort of the best that we can do for now.