Implement custom stack unwinding for Pulley (#9758)
* Implement custom stack unwinding for Pulley

The pulley calling convention may not match the native calling
convention, which is the case for s390x, so the unwinding that Wasmtime
does needs to be parameterized over what's being unwound. This commit
adds a new `Unwind` trait with various methods behind it that the
backtrace implementation is then updated to use.

* Fix alignment check for 32-bit