Add a guide for adding instructions to Pulley  (#9765)
* Add a guide for adding instructions to Pulley

In helping others contribute to Pulley I've written up the steps I took
to make this commit itself. I've picked a random test in `misc_testsuite`
that previously didn't work on Pulley and was relatively small. I
implemented a few lowerings and some new instructions and now the test
passes. I've added `pulley/CONTRIBUTING.md` to document the experience
along the way in case anyone else is interested to help out in the
future.

* Fix fuzzer build

* Review comments

* Fix pulley test

* Fix no_std build

* Fix fuzz simple push pop in miri