pulley: Ungate the `GC_TYPES` feature from Pulley
Similar to #9779 this remove the `GC_TYPES` feature from the list of
panicking features in Pulley. In doing so this then additionally fixes a
number of panics and then adds more tests that are working. Some other
minor instructions are filled out to ensure that tests are working on
both 32 and 64-bit platforms.